This PR includes 3 improvements for performance. 

## 1. Use /api/v1/dependencies

`api/v1/dependencies` returns marshal dump. `api/v1/dependencies.json` returns JSON contents. The former is faster. We should use the former.

```
marshal_url: 1.261804 sec
json_url: 2.01807 sec
```

as an example to resolve dependencies for `rake,flexmock,parallel_tests,simplecov,rr,timecop,test-unit,test-unit-rr,oj,pry,pry-nav,fluent-plugin-s3,fluent-plugin-forest,fluent-plugin-record-reformer,msgpack,json,yajl-ruby,cool.io,http_parser.rb,sigdump,tzinfo,tzinfo-data,strptime`. You can see detail with this commit https://github.com/geminabox/geminabox/pull/230/commits/41e06f716159d3849cd3c330326ba873345ffec9

## 2. Speed up gem_list_merge

```
BEFORE: 0.20931879 sec
AFTER: 0.047801906 sec
```

to resolve dependencies for `activemodel,arel,tzinfo,activerecord-deprecated_finders,i18n,multi_json,minitest,thread_safe,concurrent-ruby,method_source,term-ansicolor,treetop,gherkin,polyglot,json_pure,multi_test,cucumber-core,gherkin3,cucumber-wire,event-bus,bones,net-scp,metaclass,rspec-support`. You can see a detail with this commit https://github.com/geminabox/geminabox/commit/e53e4682669aca1c627fedf357192c118d8e5a01.

EDIT: I could improve more  using sort_by! instead of sort! https://github.com/geminabox/geminabox/pull/230/commits/39f5b2ff29be21895285c00654c73981b3750271

```
BEFORE: 1.10931879 sec
AFTER: 0.047122157
```

to resolve dependencies for
`bones,rcov,rspec,rubyforge,ZenTest,addressable,multipart-post,faraday,faraday-middleware,httpauth,jwt,multi_xml,jar-dependencies,ffi,rbnacl,curses,launchy,bones-rcov,bones-rspec,bones-rubyforge,bones-zentest,git,functional-ruby,ref,gherkin,gherkin3,trollop,term-ansicolor,hashie,oauth2,nokogiri,termios,json_pure,RubyInline,metaclass,needle,echoe,jruby-pageant,bcrypt_pbkdf,rbnacl-libsodium,minitest,psych,tins,facets,polyglot,unf_ext,sexp_processor,concurrent-ruby,method_source,jamespath`

## 3. Set Accept-Encoding: gzip

https://github.com/geminabox/geminabox/pull/230/commits/2ffcafe7f5cab3aee8c26f617d6619122db2f94a

`https://bundler.rubygems.org/api/v1/dependencies` accepts gzip encoding. Using gzip encoding was faster based on experiments. 

```
BEFORE: 1.454 sec
AFTER: 1.116 sec
```

to resolve dependencies for `unicorn,rake,server-starter,oneline_log_formatter,sinatra,builder,httpclient,nesty,faraday`