@tomlea 2016-04-22
These look like a great set of commits.

I'm no longer using geminabox, so not in a position to work on it. I've also been keeping quiet so others can takeover, but hanging on incase things get lost.

I am however in a position to give you commit rights ;)

I'll leave it in your hands, and feel free to get in touch if you need any more access (rubygems.org etc…)