@sonots 2016-04-22
@tomlea thanks, can you add me to gem author? My rubygems mail address is `sonots@gmail.com` (will remove later).