@sonots 2016-04-22
@tomlea thanks, can you add me to gem owner so that I can publish a gem? 
You can find my rubygems mail address at https://rubygems.org/profiles/sonots